+++ src/main/kotlin/mb/gitonium/GitoniumPlugin.kt
@@ -21,17 +21,36 @@ class GitoniumPlugin : Plugin<Project> {
         // Register "check for snapshot dependencies" task when publishing for project and sub-projects.
         project.afterEvaluate {
             registerCheckSnapshotDependenciesTask(this, extension)
+            registerPrintVersionTask(this)
             subprojects.forEach {
                 registerCheckSnapshotDependenciesTask(it, extension)
+                registerPrintVersionTask(it)
             }
         }
     }
 
+    /**
+     * Registers a task that checks whether any of the dependencies are snapshot dependencies when publishing.
+     *
+     * @param project The project for which to register the task.
+     * @param extension The Gitonium extension, used for the configuration.
+     */
     private fun registerCheckSnapshotDependenciesTask(project: Project, extension: GitoniumExtension) {
         if (!extension.checkSnapshotDependenciesInRelease) return
         val publishTask = project.tasks.findByName("publish") ?: return
         val checkTask = project.tasks.register<CheckSnapshotDependencies>("checkSnapshotDependencies", extension)
         publishTask.dependsOn(checkTask)
+    }
+
+    /**
+     * Registers a task that prints the current version of the project to STDOUT.
+     *
+     * @param project The project for which to register the task.
+     */
+    private fun registerPrintVersionTask(project: Project) {
+        project.tasks.register("printVersion") {
+            println(project.version)
+        }
     }
 }